Update yml ans sh

+++ push-javadoc.sh
@@ -1,23 +1,23 @@
 #!/bin/bash
 
-#if [ "$TRAVIS_REPO_SLUG" == "Wroczen/ProjektIO" ] && [ "$TRAVIS_JDK_VERSION" == "oraclejdk8" ] && [ "$TRAVIS_PULL_REQUEST" == "false" ] && [ "$TRAVIS_BRANCH" == "master" ]; then
+if [ "$TRAVIS_REPO_SLUG" == "Wyroczen/ProjektIO" ] && [ "$TRAVIS_JDK_VERSION" == "oraclejdk8" ] && [ "$TRAVIS_PULL_REQUEST" == "false" ] && [ "$TRAVIS_BRANCH" == "master" ]; then
 
-echo -e "Publishing javadoc...\n"
+  echo -e "Publishing javadoc...\n"
 
-cp -R build/docs/javadoc $HOME/javadoc-latest
+  cp -R build/docs/javadoc $HOME/javadoc-latest
 
-cd $HOME
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "travis-ci"
-git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/Wyroczen/ProjektIO gh-pages > /dev/null
+  cd $HOME
+  git config --global user.email "travis@travis-ci.org"
+  git config --global user.name "travis-ci"
+  git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/Wyroczen/ProjektIO gh-pages > /dev/null
 
-cd gh-pages
-git rm -rf ./javadoc
-cp -Rf $HOME/javadoc-latest ./javadoc
-git add -f .
-git commit -m "Latest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to gh-pages"
-git push -fq origin gh-pages > /dev/null
+  cd gh-pages
+  git rm -rf ./javadoc
+  cp -Rf $HOME/javadoc-latest ./javadoc
+  git add -f .
+  git commit -m "Latest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to gh-pages"
+  git push -fq origin gh-pages > /dev/null
 
-echo -e "Published Javadoc to gh-pages.\n"
+  echo -e "Published Javadoc to gh-pages.\n"
   
-#fi
+fi
